+++ javascript/WindowsLepik.js
@@ -18,9 +18,6 @@ class WindowsLepik extends LepikEvents_1.LepikEvents {
         this.ps = child_process_1.spawn('powershell.exe', ['-ExecutionPolicy', 'Bypass', '-File', psPath], {
             stdio: ['pipe', 'pipe', 'inherit']
         });
-        this.ps.stdout.on("data", (data) => {
-            console.log("Data:" + data.toString());
-        });
         process.on('exit', () => {
             this.ps.kill();
         });