+++ tests/singletons/getScreenSize.js
@@ -1,0 +1,14 @@
+const lepik = require("../../index.js")
+
+async function main() {
+  console.log(`Staring getScreenSize at ${Date.now()}`)
+
+  // Method implementation
+  lepik.getScreenSize().then(e => console.log(`Should print screenSize: width: ${e.width} height: ${e.height}`))
+
+
+  lepik.close()
+  console.log(`Closing getScreenSize at ${Date.now()}`)
+}
+
+main()