+++ tests/singletons/maximizeWindow.js
@@ -1,0 +1,16 @@
+const lepik = require("../../index.js")
+
+async function main() {
+  console.log(`Staring maximizeWindow at ${Date.now()}`)
+
+  // Method implementation
+  console.log("Getting id of current window")
+  let winHandle = await lepik.getActiveWindow()
+  console.log("Should maximize the active window")
+  lepik.maximizeWindow(winHandle)
+
+  lepik.close()
+  console.log(`Closing maximizeWindow at ${Date.now()}`)
+}
+
+main()