+++ tests/singletons/mouseClick.js
@@ -1,0 +1,14 @@
+const lepik = require("../../index.js")
+
+async function main() {
+  console.log(`Staring mouseClick at ${Date.now()}`)
+
+  // Method implementation
+  console.log("Click the mouse")
+  lepik.mouseClick()
+
+  lepik.close()
+  console.log(`Closing mouseClick at ${Date.now()}`)
+}
+
+main()