+++ tests/singletons/mouseMove.js
@@ -1,0 +1,14 @@
+const lepik = require("../../index.js")
+
+async function main() {
+  console.log(`Staring mouseMove at ${Date.now()}`)
+
+  // Method implementation
+  console.log("Moving the mouse to 500;500")
+  lepik.mouseDrag(500, 500)
+
+  lepik.close()
+  console.log(`Closing mouseMove at ${Date.now()}`)
+}
+
+main()